# specs/master/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: 网络小说智能问答系统 API
  description: |
    基于RAG架构的网络小说智能问答系统后端API。
    支持小说管理、智能问答、知识图谱可视化、Token统计等功能。
  version: 1.0.0
  contact:
    name: Development Team
    email: dev@novel-qa.com
servers:
  - url: http://localhost:8000/api
    description: 本地开发环境（推荐）
  - url: http://localhost:8000
    description: 本地开发环境（不带前缀，已废弃）
...
paths:
...
  /graph/relations/{novelId}:
    get:
      tags: [graph]
      summary: 获取角色关系图数据
      parameters:
        - $ref: '#/components/parameters/NovelId'
        - name: startChapter
          in: query
          schema:
            type: integer
          description: 起始章节（可选，用于查看特定章节范围的关系）
        - name: endChapter
          in: query
          schema:
            type: integer
          description: 结束章节（可选）
      responses:
        '200':
          description: 成功返回关系图数据
          content:
            application/json:
              schema:
                type: object
                properties:
                  nodes:
                    type: array
                    items:
                      $ref: '#/components/schemas/GraphNode'
                  edges:
                    type: array
                    items:
                      $ref: '#/components/schemas/GraphEdge'
  /graph/relations/{novelId}/node/{nodeId}:
    get:
      tags: [graph]
      summary: 获取节点详细信息
      parameters:
        - name: novelId
          in: path
          required: true
          schema:
            type: integer
          description: 小说ID
        - name: nodeId
          in: path
          required: true
          schema:
            type: string
          description: 节点ID（实体名称）
      responses:
        '200':
          description: 节点详细信息
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  type:
                    type: string
                  attributes:
                    type: object
                  neighbors:
                    type: array
                    items:
                      type: object
        '404':
          $ref: '#/components/responses/NotFound'
  /graph/timeline/{novelId}:
    get:
      tags: [graph]
      summary: 获取时间线数据
      parameters:
        - $ref: '#/components/parameters/NovelId'
      responses:
        '200':
          description: 成功返回时间线数据
          content:
            application/json:
              schema:
                type: object
                properties:
                  events:
                    type: array
                    items:
                      $ref: '#/components/schemas/TimelineEvent'
...
  /config/test-connection:
    post:
      tags: [config]
      summary: 测试智谱AI连接
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [apiKey]
              properties:
                apiKey:
                  type: string
                  description: 智谱AI API Key
      responses:
        '200':
          description: 连接测试结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  model_tested:
                    type: string
  /config:
    get:
      tags: [config]
      summary: 获取系统配置
      responses:
        '200':
          description: 成功返回配置
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemConfig'
    
    put:
      tags: [config]
      summary: 更新系统配置
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SystemConfig'
      responses:
        '200':
          description: 配置更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean

  # ==================== 统计分析 ====================
...
components:
  parameters:
    NovelId:
      name: novelId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
      description: 小说ID
  responses:
...
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
...
  schemas:
...
    GraphNode:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [character, location, organization, item]
        importance:
          type: number
          minimum: 0
          maximum: 1
    GraphEdge:
      type: object
      properties:
        source:
          type: string
        target:
          type: string
        relationType:
          type: string
        startChapter:
          type: integer
        endChapter:
          type: integer
          nullable: true
    TimelineEvent:
      type: object
      properties:
        chapterNum:
          type: integer
        narrativeOrder:
          type: integer
        actualTime:
          type: string
        description:
          type: string
...
    SystemConfig:
      type: object
      properties:
        defaultModel:
          type: string
          enum: [glm-4-flash, glm-4, glm-4-plus, glm-4-5, glm-4-6]
        topK:
          type: integer
          minimum: 10
          maximum: 50
          default: 30
        chunkSize:
          type: integer
          minimum: 400
          maximum: 800
          default: 550
        chunkOverlap:
          type: integer
          minimum: 50
          maximum: 200
          default: 125
        enableSelfRAG:
          type: boolean
          default: true
        enableSmartRouting:
          type: boolean
          default: true

    # ========== 统计相关 ==========
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: JWT认证（可选，个人应用可不启用）

# 默认不启用安全认证（个人应用）
# security:
#   - BearerAuth: []